Stop throwing away company info
Company information is returned as a top-level key in the response to `/account/who_am_i`. By feeding that response directly into `Harvest::User.parse`, we throw away that data. This change preserves that data and makes it accessible by simply tacking it on to the returned user object as the `company` attribute.